+++ Source/Core/Prune/DependencyEvaluator.cs
@@ -56,21 +56,4 @@ namespace Microsoft.Boogie
       this.declaration = declaration;
     }
   }
-
-  class ConstantVisitor : DependencyEvaluator
-  {
-    protected ConstantVisitor(Declaration declaration) : base(declaration)
-    {
-    }
-    
-    public static DependencyEvaluator GetDependencies(Constant constant)
-    {
-      var result = new ConstantVisitor(constant);
-      foreach (var definitionAxiom in constant.DefinitionAxioms) {
-        result.AddOutgoing(definitionAxiom);
-      }
-      result.Visit(constant);
-      return result;
-    }
-  }
 }
